Corrected some bugs
I have added the support of the colors on the messages, added a
config.yml file with currently one config (enable or disable the Action
Bar for display messages), corrected a glitch which allowed to send
empty messages. But I have a problem for create the locale.yml file
which will contains the messages of the plugin for translating.